Fully upgraded plugin to 1.14.4 Spigot and tidyed up code
