+++ FirstPlugin/src/main/java/me/tyler/firstplugin/FirstPlugin.java
@@ -1,16 +1,13 @@
 package me.tyler.firstplugin;
 
-import me.tyler.firstplugin.onJoin.onJoin;
-import me.tyler.firstplugin.onJoin.onLeave;
-import org.bukkit.command.Command;
-import org.bukkit.command.CommandSender;
+import me.tyler.firstplugin.commands.DieCommand;
+import me.tyler.firstplugin.commands.HpCommand;
+import me.tyler.firstplugin.events.OnJoin;
+import me.tyler.firstplugin.events.OnLeave;
 import org.bukkit.entity.Player;
-import org.bukkit.event.Event;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
-import org.bukkit.event.block.BlockPlaceEvent;
 import org.bukkit.event.player.PlayerBedLeaveEvent;
-import org.bukkit.event.player.PlayerMoveEvent;
 import org.bukkit.plugin.java.JavaPlugin;
 
 public final class FirstPlugin extends JavaPlugin implements Listener {
@@ -20,8 +17,12 @@ public final class FirstPlugin extends JavaPlugin implements Listener {
         // Plugin startup logic
         System.out.println("Your First Plugin Has been Enabled!");
         getServer().getPluginManager().registerEvents(this, this);
-        getServer().getPluginManager().registerEvents(new onJoin(), this);
-        getServer().getPluginManager().registerEvents(new onLeave(), this);
+        getServer().getPluginManager().registerEvents(new OnJoin(), this);
+        getServer().getPluginManager().registerEvents(new OnLeave(), this);
+
+        // Register commands
+        this.getCommand("die").setExecutor(new DieCommand());
+        this.getCommand("hp").setExecutor(new HpCommand());
     }
 
     @Override
@@ -36,6 +37,10 @@ public final class FirstPlugin extends JavaPlugin implements Listener {
         player.sendMessage("You left the bed");
     }
 
+    // Removed command handling here and moved to seperate classes:
+    // me.tyler.firstplugin.commands.*
+
+    /*
     @Override
     public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
         if(command.getName().equals("die")){
@@ -43,7 +48,7 @@ public final class FirstPlugin extends JavaPlugin implements Listener {
                 Player player = (Player) sender;
                 player.sendMessage("Die Activated...");
                 player.setHealth(0);
-            }else{
+            } else {
                 System.out.println("You need to be a player to execute this command!");
             }
         }
@@ -53,12 +58,12 @@ public final class FirstPlugin extends JavaPlugin implements Listener {
                 Player player = (Player) sender;
                 player.sendMessage("HeathPls Activated...");
                 player.setHealth(20);
-            }else{
+            } else {
                 System.out.println("You need to be a player to execute this command!");
             }
         }
 
         return false;
     }
-
+    */
 }
