+++ FirstPlugin/src/main/java/me/tyler/firstplugin/events/OnJoin.java
@@ -1,0 +1,16 @@
+package me.tyler.firstplugin.events;
+
+import org.bukkit.ChatColor;
+import org.bukkit.entity.Player;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.Listener;
+import org.bukkit.event.player.PlayerJoinEvent;
+
+public class OnJoin implements Listener {
+    @EventHandler
+    void onPlayerJoin(PlayerJoinEvent event){
+        Player player = event.getPlayer();
+        event.setJoinMessage(ChatColor.AQUA + "Welcome back " + ChatColor.GOLD + player.getDisplayName() + ChatColor.AQUA + "!");
+
+    }
+}
