+++ FirstPlugin/src/main/java/me/tyler/firstplugin/events/OnLeave.java
@@ -1,0 +1,15 @@
+package me.tyler.firstplugin.events;
+
+import org.bukkit.ChatColor;
+import org.bukkit.entity.Player;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.Listener;
+import org.bukkit.event.player.PlayerQuitEvent;
+
+public class OnLeave implements Listener {
+    @EventHandler
+    void onPlayerQuit(PlayerQuitEvent event) {
+        Player player = event.getPlayer();
+        event.setQuitMessage(ChatColor.AQUA + "Good Bye " + ChatColor.GOLD + player.getDisplayName() + ChatColor.AQUA +"!");
+    }
+}
