--- FirstPlugin/src/main/java/me/tyler/firstplugin/onJoin/onJoin.java
@@ -1,18 +1,0 @@
-package me.tyler.firstplugin.onJoin;
-
-import org.bukkit.ChatColor;
-import org.bukkit.entity.Player;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.Listener;
-import org.bukkit.event.player.PlayerJoinEvent;
-
-public class onJoin implements Listener {
-
-    @EventHandler
-    void onPlayerJoin(PlayerJoinEvent event){
-        Player player = event.getPlayer();
-        event.setJoinMessage(ChatColor.AQUA + "Welcome back " + ChatColor.GOLD + player.getDisplayName() + ChatColor.AQUA + "!");
-
-    }
-
-}
