--- FirstPlugin/src/main/java/me/tyler/firstplugin/onJoin/onLeave.java
@@ -1,17 +1,0 @@
-package me.tyler.firstplugin.onJoin;
-
-import org.bukkit.ChatColor;
-import org.bukkit.entity.Player;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.Listener;
-import org.bukkit.event.player.PlayerQuitEvent;
-
-public class onLeave implements Listener {
-
-    @EventHandler
-    void onPlayerQuit(PlayerQuitEvent event) {
-        Player player = event.getPlayer();
-        event.setQuitMessage(ChatColor.AQUA + "Good Bye " + ChatColor.GOLD + player.getDisplayName() + ChatColor.AQUA +"!");
-    }
-
-}
